refactor(vscode): share diff review notice and setup
